# services/gateway/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OlimpicApp — Tournament Platform API
  description: |
    Multi-sport tournament management platform.
    All endpoints require JWT authentication via httpOnly cookies unless marked as public.
    The gateway proxies requests to internal microservices.
  version: 1.0.0
  contact:
    name: Tournament Platform Team
servers:
  - url: http://localhost:3000
    description: Local development gateway
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    Venue:
      type: object
      properties:
        id: { type: string, format: uuid }
        tournamentId: { type: string, format: uuid, nullable: true }
        name: { type: string }
        address: { type: string, nullable: true }
        locationUrl: { type: string, nullable: true }
        capacity: { type: integer, nullable: true }
        surfaceType: { type: string, nullable: true }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
...
security:
  - cookieAuth: []
paths:
...
  /api/tournaments:
    get:
      tags: [Tournaments]
      summary: List tournaments
      parameters:
        - name: sportId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { type: string }
        - name: season
          in: query
          schema: { type: string }
      responses:
        '200': { description: Tournament list }
    post:
      tags: [Tournaments]
      summary: Create a tournament (organizer+)
      responses:
        '201': { description: Tournament created }
  /api/tournaments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Tournaments]
      summary: Get tournament by ID
      responses:
        '200': { description: Tournament details }
    put:
      tags: [Tournaments]
      summary: Update tournament (organizer+)
      responses:
        '200': { description: Updated }
    delete:
      tags: [Tournaments]
      summary: Delete tournament (admin)
      responses:
        '204': { description: Deleted }

  # ── Teams ────────────────────────────────────────
...
  /api/venues:
    get:
      tags: [Venues]
      summary: List venues
      parameters:
        - name: tournamentId
          in: query
          schema: { type: string, format: uuid }
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Venue list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Venue' }
    post:
      tags: [Venues]
      summary: Create venue (organizer+)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                tournamentId: { type: string, format: uuid, nullable: true }
                name: { type: string, minLength: 2, maxLength: 200 }
                address: { type: string, nullable: true }
                locationUrl: { type: string, format: uri, nullable: true }
                capacity: { type: integer, minimum: 1, nullable: true }
                surfaceType: { type: string, nullable: true }
      responses:
        '201': { description: Venue created }
...